- Maven Projekt erstellt. - Encoding angepasst auf UTF-8 - Dateien an die richtige Position verschoben.
