--- module-info.java
@@ -1,8 +1,0 @@
-module Schulden {
-	requires java.desktop;
-	requires javafx.controls;
-	requires javafx.graphics;
-	requires javafx.base;
-
-	opens schulden.list to javafx.controls, javafx.base, javafx.graphics;
-}
