+++ src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module Schulden {
+	requires java.desktop;
+	requires javafx.controls;
+	requires javafx.graphics;
+	requires javafx.base;
+
+	exports schulden.list;
+	opens schulden.list to javafx.controls, javafx.base, javafx.graphics;
+}
